+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for directory-contents
 
+## 0.2.0.0
+
+* Add `Data.Data` instance for `DirTree` and `Symlink`
+* Add `System.Directory.Contents.Zipper` for convenient navigation and modification of directory contents
+
 ## 0.1.0.0
 
 * Build recursive directory trees and try to avoid symlink cycles. Includes functions to walk a directory tree, filter it, and display it.